+++ app/BP_admin/templates/settings.html
@@ -236,5 +236,5 @@
 
 
 {% block page_script %}
-<script src="{{url_for('BP_admin.static', filename='js/settings.js')}}"></script>
+<script src="{{url_for('static', filename='js/account-form-validators.js')}}"></script>
 {% endblock page_script %}
